Move `fips_is_prime()` to `fips` module
